# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Odoo POS Cafe API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /payment-methods:
    get:
      operationId: listPaymentMethods
      tags: [paymentMethods]
      summary: List payment methods
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Payment methods list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PaymentMethod"
  /payment-methods/{id}:
    put:
      operationId: updatePaymentMethod
      tags: [paymentMethods]
      summary: Update payment method
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePaymentMethodBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentMethod"

  # Floors
...
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List orders
      security:
        - bearerAuth: []
      parameters:
        - name: sessionId
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [DRAFT, PAID]
        - name: customerId
          in: query
          schema:
            type: string
        - name: dateFrom
          in: query
          schema:
            type: string
        - name: dateTo
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Orders list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OrderSummary"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create order
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderBody"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Order detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
    delete:
      operationId: deleteOrder
      tags: [orders]
      summary: Delete draft order
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /orders/{id}/status:
    put:
      operationId: updateOrderStatus
      tags: [orders]
      summary: Update order status
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrderStatusBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}/lines:
    post:
      operationId: addOrderLine
      tags: [orders]
      summary: Add line to order
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddOrderLineBody"
      responses:
        "201":
          description: Line added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderLine"
  /orders/{id}/lines/{lineId}:
    put:
      operationId: updateOrderLine
      tags: [orders]
      summary: Update order line
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: lineId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrderLineBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderLine"
    delete:
      operationId: deleteOrderLine
      tags: [orders]
      summary: Remove order line
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: lineId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /orders/table/{tableId}:
    get:
      operationId: getTableOrder
      tags: [orders]
      summary: Get active order for table
      security:
        - bearerAuth: []
      parameters:
        - name: tableId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Active order for table (or null)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  # Payments
  /payments:
    get:
      operationId: listPayments
      tags: [payments]
      summary: List payments
      security:
        - bearerAuth: []
      parameters:
        - name: sessionId
          in: query
          schema:
            type: string
        - name: dateFrom
          in: query
          schema:
            type: string
        - name: dateTo
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Payments list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Payment"
    post:
      operationId: createPayment
      tags: [payments]
      summary: Create payment
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePaymentBody"
      responses:
        "201":
          description: Payment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"

  # Kitchen
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ProductCategory:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        color:
          type: string
      required:
        - id
        - name
        - color
...
    ProductVariant:
      type: object
      properties:
        id:
          type: string
        productId:
          type: string
        attribute:
          type: string
        value:
          type: string
        extraPrice:
          type: number
        unit:
          type: string
      required:
        - id
        - productId
        - attribute
        - value
        - extraPrice
        - unit
...
    Product:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        categoryId:
          type: string
        category:
          $ref: "#/components/schemas/ProductCategory"
        price:
          type: number
        unit:
          type: string
        tax:
          type: number
        description:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        variants:
          type: array
          items:
            $ref: "#/components/schemas/ProductVariant"
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - categoryId
        - price
        - unit
        - tax
        - createdAt
...
    PaymentMethod:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
          enum: [CASH, DIGITAL, UPI]
        enabled:
          type: boolean
        upiId:
          type: ["string", "null"]
      required:
        - id
        - name
        - enabled
    UpdatePaymentMethodBody:
      type: object
      properties:
        enabled:
          type: boolean
        upiId:
          type: ["string", "null"]
...
    Table:
      type: object
      properties:
        id:
          type: string
        floorId:
          type: string
        tableNumber:
          type: string
        seats:
          type: integer
        active:
          type: boolean
        token:
          type: string
      required:
        - id
        - floorId
        - tableNumber
        - seats
        - active
        - token
...
    Customer:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        state:
          type: ["string", "null"]
        country:
          type: string
      required:
        - id
        - name
        - country
...
    OrderLine:
      type: object
      properties:
        id:
          type: string
        orderId:
          type: string
        productId:
          type: string
        product:
          $ref: "#/components/schemas/Product"
        qty:
          type: integer
        unitPrice:
          type: number
        tax:
          type: number
        subTotal:
          type: number
        total:
          type: number
        note:
          type: ["string", "null"]
      required:
        - id
        - orderId
        - productId
        - qty
        - unitPrice
        - tax
        - subTotal
        - total
    Order:
      type: object
      properties:
        id:
          type: string
        sessionId:
          type: string
        tableId:
          type: ["string", "null"]
        table:
          $ref: "#/components/schemas/Table"
        customerId:
          type: ["string", "null"]
        customer:
          $ref: "#/components/schemas/Customer"
        status:
          type: string
          enum: [DRAFT, PAID]
        createdAt:
          type: string
          format: date-time
        lines:
          type: array
          items:
            $ref: "#/components/schemas/OrderLine"
        payments:
          type: array
          items:
            $ref: "#/components/schemas/Payment"
      required:
        - id
        - sessionId
        - status
        - createdAt
    OrderSummary:
      type: object
      properties:
        id:
          type: string
        sessionId:
          type: string
        tableId:
          type: ["string", "null"]
        table:
          $ref: "#/components/schemas/Table"
        customerId:
          type: ["string", "null"]
        customer:
          $ref: "#/components/schemas/Customer"
        status:
          type: string
          enum: [DRAFT, PAID]
        createdAt:
          type: string
          format: date-time
        total:
          type: number
      required:
        - id
        - sessionId
        - status
        - createdAt
        - total
...
    CreateOrderBody:
      type: object
      properties:
        sessionId:
          type: string
        tableId:
          type: ["string", "null"]
        customerId:
          type: ["string", "null"]
      required:
        - sessionId
    UpdateOrderStatusBody:
      type: object
      properties:
        status:
          type: string
          enum: [DRAFT, PAID]
      required:
        - status
    AddOrderLineBody:
      type: object
      properties:
        productId:
          type: string
        qty:
          type: integer
          default: 1
        note:
          type: ["string", "null"]
      required:
        - productId
    UpdateOrderLineBody:
      type: object
      properties:
        qty:
          type: integer
        note:
          type: ["string", "null"]
    Payment:
      type: object
      properties:
        id:
          type: string
        orderId:
          type: string
        method:
          type: string
          enum: [CASH, DIGITAL, UPI]
        amount:
          type: number
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - orderId
        - method
        - amount
        - createdAt
    CreatePaymentBody:
      type: object
      properties:
        orderId:
          type: string
        method:
          type: string
          enum: [CASH, DIGITAL, UPI]
        amount:
          type: number
      required:
        - orderId
        - method
        - amount
...