# lib/api-zod/src/generated/types/updateOrderLineBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Odoo POS Cafe API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateOrderLineBody {
  qty?: number;
  /** @nullable */
  note?: string | null;
}